Enhance ECV filters with nested groups and improve name overrides
Updates the Entity Collection View filters to support recursive boolean logic (AND/OR/NOT groups) and adds a new distance constraint. Name overrides are expanded to support items by dynamically substituting localized string arguments for clilocs.

Additional changes:
- Adds a "Target Owner" context menu action.
- Optimizes attack commands by caching the last targeted combatant to reduce redundant packets.
- Improves UI consistency by grouping options in the Friends and Trap Pouch tabs.
- Fixes assembly resolution and improves extension initialization error handling.

+++ ClassicAssist/UO/Commands.cs
@@ -254,7 +254,7 @@ namespace ClassicAssist.UO
                     serial = ( packet[7] << 24 ) | ( packet[8] << 16 ) | ( packet[9] << 8 ) | packet[10];
                     x = ( packet[11] << 8 ) | packet[12];
                     y = ( packet[13] << 8 ) | packet[14];
-                    z = ( packet[15] << 8 ) | packet[16];
+                    z = (short) ( ( packet[15] << 8 ) | packet[16] );
                     itemID = ( packet[17] << 8 ) | packet[18];
 
                     are.Set();
